Speed up Javascript commit line highlighter
Each line of a commit diff has a unique ID. Previously, we'd use
'document.querySelector' when searching for lines in the DOM as part of
the line highliter. Now, we additionally annotate these DOM elements
with an 'id="<unique id>"' so we can use 'document.getElementById',
which is much faster.