Whole-site password protection #88
It's now possible to require HTTP authentication for all of the Web
interface (not only push/pull). This isn't available from the command
line.

Note that when both whole-site password protection and Smart HTTP mode
is enabled, push is allowed by default. It can be disallowed using the
'disable_push' option.