Allow trailing slash for history and submodule views.
Fixes #190. If an URL rule in flask does not end with a part with a path
converter it is better to end such rule with a slash, because
documentation at
http://flask.pocoo.org/docs/0.12/api/#url-route-registrations says:

"1. If a rule ends with a slash and is requested without a slash by the
user, the user is automatically redirected to the same page with a
trailing slash attached.
2. And if a rule does not end with a trailing slash and the user
requests the page with a trailing slash, a 404 not found is raised."