Recognize lower-case README file names
Simply converts all known readme naming conventions and file names
within the tree to lower case on file discovery. This ensures double the
coverage of acceptable readme filenames.

This patch is the result of a discussion over making the known readme
naming conventions extendable by the user on the Klaus mailing list.
[ See https://groups.google.com/forum/#!topic/klaus-users/mXjDxyqbeMQ ]