Support nested namespaces (without ~prefix)
This is a major change that changes all existing route fomats.
The base route of each repo is no longer `/repo` or `/~ns/repo` but
`/repo/-/` or `/deeply/nested/repo/-/`, whereas the Dulwich/SMARTHTTP
routes are at `/repo.git` or `/deeply/nested/repo.git`.

Klaus is expected to be used together with some method of hosting Git
repositories, such as plain SSH access, git-shell, or e.g. gitolite.
All of these allow organizing the repositories freely, creating directory
structures (and even mixing with other files) as wanted.

This routing scheme is inspired by GitLab, where repositories and groups
can be nested in this way also. It allows klaus to have that same flexibility
and keeps the SMARTHTTP clone URL "symmetric" to what e.g. an SSH clone URL may
look like:

    git@host:path/to/repo.git
    http://host/path/to/repo.git

Some might find the `-` in the URL irritating. The biggest issue is simply
that changing the URL scheme is not backwards compatible.

It would make sense to redirect `/repo` to `/repo/-/` automatically,
which would solve the most common issue (links to repositories not working).